+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADEntitlementManagementAccessPackage/MSFT_AADEntitlementManagementAccessPackage.psm1
@@ -111,7 +111,7 @@ function Get-TargetResource
                 -ErrorAction SilentlyContinue
         }
 
-        if ($null -eq $getValue)
+        if ($null -eq $getValue -and -not [System.String]::IsNullOrEmpty($id))
         {
             Write-Verbose -Message "Nothing with id {$id} was found"
 
@@ -783,10 +783,10 @@ function Test-TargetResource
     $CurrentValues = Get-TargetResource @PSBoundParameters
     $ValuesToCheck = ([Hashtable]$PSBoundParameters).clone()
 
-    if ($CurrentValues.Ensure -eq 'Absent')
+    if ($CurrentValues.Ensure -eq 'Absent' -and $Ensure -eq 'Absent')
     {
-        Write-Verbose -Message "Test-TargetResource returned $false"
-        return $false
+        Write-Verbose -Message "Test-TargetResource returned $true"
+        return $true
     }
     $testResult = $true
 
@@ -801,6 +801,11 @@ function Test-TargetResource
             foreach ($s in [Array]$source)
             {
                 $s.remove('Id')
+            }
+
+            if ($target.getType().Name -like '*CimInstance*')
+            {
+                $target = Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $target
             }
             foreach ($t in [Array]$target)
             {